Refactored the authorship pallet
