+++ runtime/src/configs/mod.rs
@@ -1,3 +1,4 @@
+pub mod authorship;
 pub mod timestamp;
 pub mod balances;
 pub mod transaction_payment;
